fix - reg - Fixed GradientRot not advancing to the next screen
We've fixed GradientRot not advancing to the next screen

---

Because of the missing floating-point casts caused by the conversion from Visual Basic, we've made the color steps really consist of double-precision floating point numbers, thus actually letting GradientRot advance to the next color.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1

+++ public/Nitrocid/Misc/Screensaver/Displays/GradientRot.cs
@@ -313,9 +313,9 @@ namespace KS.Misc.Screensaver.Displays
             int RampColorRedThreshold = RedColorNumFrom - RedColorNumTo;
             int RampColorGreenThreshold = GreenColorNumFrom - GreenColorNumTo;
             int RampColorBlueThreshold = BlueColorNumFrom - BlueColorNumTo;
-            double RampColorRedSteps = RampColorRedThreshold / RampFrameSpaces;
-            double RampColorGreenSteps = RampColorGreenThreshold / RampFrameSpaces;
-            double RampColorBlueSteps = RampColorBlueThreshold / RampFrameSpaces;
+            double RampColorRedSteps = RampColorRedThreshold / (double)RampFrameSpaces;
+            double RampColorGreenSteps = RampColorGreenThreshold / (double)RampFrameSpaces;
+            double RampColorBlueSteps = RampColorBlueThreshold / (double)RampFrameSpaces;
             DebugWriter.WriteDebugConditional(Screensaver.ScreensaverDebug, DebugLevel.I, "Set thresholds (RGB: {0};{1};{2})", RampColorRedThreshold, RampColorGreenThreshold, RampColorBlueThreshold);
             DebugWriter.WriteDebugConditional(Screensaver.ScreensaverDebug, DebugLevel.I, "Steps by {0} spaces (RGB: {1};{2};{3})", RampFrameSpaces, RampColorRedSteps, RampColorGreenSteps, RampColorBlueSteps);
